rename component and support basic visibility
